Module Sihl_user__Authn.Make

Parameters

SessionService : Sihl_contract.Session.Sig

UserService : Sihl_contract.User.Sig

Signature

include Sihl_core.Container.Service.Sig

val lifecycle : Sihl_core.Container.Lifecycle.t

val find_user_in_session_opt : Sihl_type.Session.t -> Sihl_type.User.t option Lwt.t

Find currently logged in user in the current context.

Make sure to call authenticate_session before or apply the required session and authentication middlewares.

val find_user_in_session : Sihl_type.Session.t -> Sihl_type.User.t Lwt.t

Find currently logged in user in the current context.

Make sure to call authenticate_session before or apply the required session and authentication middlewares.

val authenticate_session : Sihl_type.User.t -> Sihl_type.Session.t -> unit Lwt.t

Assign a user to the current anonymous session.

Use authenticate_session ctx user to log in a user. If a user is already assigned to session, replace the user.

val unauthenticate_session : Sihl_type.Session.t -> unit Lwt.t

Log user out.

Remove user from current session so that session is anonymous again.

val register : unit -> Sihl_core.Container.Service.t